backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
